# src/components/docs/docSwagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is WEBSTORE API REQUIREMENT OF TRONTRON",
    "version": "1.0.5",
    "title": "Swagger WebStore",
    "contact": {
      "email": "lenguyenyenhong1199@gmail.com"
    }
  },
  "host": "http://localhost:5000",
  "basePath":"/",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/product/":{
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "Add new Product",
        "description": "",
        "operationId": "createNewProduct",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          
            {
              "name": "productName",
              "in": "formData",
              "description": " Name of the Product",
              "required": true,
              "type": "string"
            },
            {
              "name": "price",
              "in": "formData",
              "description": "price of the Product",
              "required": true,
              "type": "integer"
            },
            {
              "name": "amount",
              "in": "formData",
              "description": "amount of the Product",
              "required": true,
              "type": "string"
            },
            {
              "name": "description",
              "in": "formData",
              "description": "description of the Product",
              "required": true,
              "type": "string"
            },
            {
              "name": "distributor",
              "in": "formData",
              "description": "distributor of the Product",
              "required": true,
              "type": "string"
            },
            {
              "name": "categoryId",
              "in": "formData",
              "description": "categoryId of the Product",
              "required": true,
              "type": "string",
              "ref":"#/definitions/Category"
            },
            {
              "name": "images",
              "in": "formData",
              "description": "images of the Product",
              "required": true,
              "type": "array",
              "items":{
                "type":"file"
              }
            }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  
                    "$ref": "#/definitions/Product"             
                  
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/product/{productId}":{
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Find Product by ID",
        "description": "Returns a single Product",
        "operationId": "getProductById",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  
                    "$ref": "#/definitions/Product"             
                  
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "api_key": [
              
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "Product"
        ],
        "summary": "Updates an Product in the store with form data",
        "description": "",
        "operationId": "updateProductById",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user that needs to be updated",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "productName",
            "in": "formData",
            "description": " Name of the Product",
            "required": false,
            "type": "string"
          },
          {
            "name": "price",
            "in": "formData",
            "description": "price of the Product",
            "required": false,
            "type": "integer"
          },
          {
            "name": "amount",
            "in": "formData",
            "description": "amount of the Product",
            "required": false,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "description of the Product",
            "required": false,
            "type": "string"
          },
          {
            "name": "distributor",
            "in": "formData",
            "description": "distributor of the Product",
            "required": false,
            "type": "string"
          },
          {
            "name": "categoryId",
            "in": "formData",
            "description": "categoryId of the Product",
            "required": false,
            "type": "string",
            "ref":"#/definitions/Category"
          },
          {
            "name": "images",
            "in": "formData",
            "description": "images of the Product",
            "required": false,
            "type": "array",
            "items":{
              "type":"file"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  
                    "$ref": "#/definitions/Product"             
                  
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "summary": "Deletes an Product",
        "description": "Delete Product By Id User",
        "operationId": "deteleProductById",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "categoryId",
            "in": "path",
            "description": "categoryId to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{
                  "type": "string",
                  "example":"Delete Successfully"                 
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/product/?keyName={name}":{
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Search Product By Name",
        "description": "",
        "operationId": "searchProductByName",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [          
            {
              "name": "keyName",
              "in": "query",
              "description": "Key Name of the Product",
              "required": true,
              "type": "string"
            }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"200"
                },
                "success":{
                  "type":"boolean",
                  "example":"true"
                },
                "data":{                  
                    "type":"array",
                    "item":{
                      "type":"object",
                      "properties":{
                        "productName":{
                          "type":"string",
                          "example":"Son A"
                        }
                      }
                    }
                  
                }
              }
            }
          },
          "400":{
            "description":"Error Validate",
            "schema": {
              "type":"object",
              "properties":{
                "code": {
                  "type":"integer",
                  "example":"400"
                },
                "success":{
                  "type":"boolean",
                  "example":"false"
                },
                "data":{
                  "type": "string",
                  "example":"Validate Error"
                  }
                }
              }
            },
         "500":{
          "description":"Server error",
          "schema": {
            "type":"object",
            "properties":{
              "code": {
                "type":"integer",
                "example":"500"
              },
              "success":{
                "type":"boolean",
                "example":"false"
              },
              "data":{
                "type": "string",
                "example":"Server error"
                }
              }
            }
         }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read:pets": "read your pets",
        "write:pets": "modify pets in your account"
      }
    }
  },
  "definitions": {
...
    "Category": {
      "type": "object",
      "properties": {
        "categoryName": {
          "type": "string"
        },
        "descriptionCategory": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Category"
      }
    },
    "Product": {
      "type": "object",
       "properties": {
        "productName": {
          "type": "integer",
          "format": "int64"
        },
        "price": {
          "type":"integer"
          
        },
        "amount": {
          "type": "string"
          
        },
        "categoryId":{
          "type":"string",
          "ref":"#/definitions/Category"
        },
        "description":{
          "type":"string"
        },
        "distributor":{
          "type":"string"
        },
        "images":{
          "type":"array",
          "items":{
            "type":"file",
            "example":"imagename.jpg"
          }
        },
        "remark":{
          "type":"boolean"
        }
        
       
      },
      "xml": {
        "name": "Product"
      }
    },
...
  },
...
}